Fixed some build issues.
git-svn-id: https://svn.code.sf.net/p/miktex/code/miktex/trunk@1525 3fc76331-e2db-4b81-af9c-3600f57961af

+++ BuildUtilities/c4p/main.cpp
@@ -67,7 +67,7 @@ unsigned extra_indent;
 unsigned max_lines_per_c_file;
 
 int yyparse ();
-extern long yylineno;
+extern int yylineno;
 extern _TCHAR * yytext;
 
 /* _________________________________________________________________________
